### 2023-04-11 #585 @derek-corcoran-barrios
Presubmission of the AICcPermanova R Package
Submitting Author Name: Derek Corcoran
Submitting Author Github Handle: @derek-corcoran-barrios

Repository:  https://github.com/Sustainscapes/AICcPerm
Submission type: Pre-submission

---

```
Package: AICcPermanova
Title: Model Selection of PERMANOVA Models Using AICc
Version: 0.0.3
    person(given = "Derek", family ="Corcoran",
    email = "derek.corcoran.barrios@gmail.com", role = c("aut", "cre"))
Description: Provides tools for model selection and model averaging of PerMANOVA
    models using Akaike Information Criterion corrected for small sample sizes
    (AICc) and Information Theoretic criteria principles. The package is built
    around the PERMANOVA analysis from the 'vegan' package and provides a
    streamlined workflow for generating and comparing models, obtaining model
    weights, and summarizing results using model averaging approaches.  The
    methods implemented in this package are based on the practical information-
    theoretic approach described by Burnham, K. P. and Anderson, D. R. (2002)
    (<doi:10.1007/b97636>).
    broom,
    car,
    doParallel,
    foreach,
    furrr,
    future,
    parallel,
    vegan
    https://sustainscapes.github.io/AICcPerm/
    covr,

```

The package I build is built to generate a clear workflow for community analysis, specifically for model selection in Permanova, which are a form of regression

Not yet, I am looking into doing it right now, I expect to have it done by the end of this week

Community ecologist in general, I see it as very useful for microbiologists working with dna barcoding taxonomy, botanist that do plots for community analysis, etc.

This package is based on the vegan package, so that package has the analysis made here, however it does not have the model selection capabilities.

There are 2 issues largely for me. I would like to know if the categories I am choosing for the package are the right ones, and the second is that I would like this package to be in either the Methods in ecology and evolution or The Journal of Open Source Software, however when I am asked to check that in the last part of the submission, I am not completely done with the writing of the article, would this be needed to check those boxes?
