### 2023-04-11 16c0979
Fixed the akaike adjusted rsq function and added the citation
