+++ tests/testthat/test-akaike_adjusted_rsq.R
@@ -1,5 +1,6 @@
 # Define a sample input data frame
 df <- data.frame(
+  form = c("Distance ~ A1", "Distance ~ A2", "Distance ~ A3"),
   A1 = c(1, 2, 3), A2 = c(4, 5, NA), A3 = c(6, 7, 8),
   AICc = c(10, 11, 12),
   max_vif = c(1.5, 2.1, 4.9), k = c(3, 4, 5),
