+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   getting-started-with-aiccpermanova: getting-started-with-aiccpermanova.html
-last_built: 2023-04-11T08:16Z
+last_built: 2023-04-11T19:38Z
 urls:
   reference: https://sustainscapes.github.io/AICcPermanova/reference
   article: https://sustainscapes.github.io/AICcPermanova/articles
