### 2023-04-23 #585 @maurolepore
Dear @derek-corcoran-barrios

Sorry for the delay.

Answering your questions:

1. Does the package do statistics itself of rather wrap other packages that do the stats? If it doesn't do stats itself then the standard (non-stats) category "workflow automation" seems like a good fit.

2. In the submission template I see "Although not required, we strongly recommend having a full manuscript prepared when you submit here." Because it's not required, I think it's OK for you to check that box, maybe with a comment on when you think it'll be ready?
