### 2023-04-24 #585 @derek-corcoran-barrios
1. This sound great to me, just to answer your question most of the
   stats except for the AICc is calculated by vegan or CAR, so I guess
   workflow automation fits
   2. I will do that

How should I proceed from here? Do I make a new issue?

Cheers

Derek

On Sun, Apr 23, 2023 at 11:13 PM Mauro Lepore ***@***.***>
wrote:

> Dear @derek-corcoran-barrios <https://github.com/derek-corcoran-barrios>
>
> Sorry for the delay.
>
> Answering your questions:
>
>    1.
>
>    Does the package do statistics itself of rather wrap other packages
>    that do the stats? If it doesn't do stats itself then the standard
>    (non-stats) category "workflow automation" seems like a good fit.
>    2.
>
>    In the submission template I see "Although not required, we strongly
>    recommend having a full manuscript prepared when you submit here." Because
>    it's not required, I think it's OK for you to check that box, maybe with a
>    comment on when you think it'll be ready?
>
