### 2023-04-24 #585 @derek-corcoran-barrios
Thanks I will do that

Cheers

On Mon, Apr 24, 2023 at 12:14 PM Mauro Lepore ***@***.***>
wrote:

> Great.
>
> Yes, please go ahead and open a new issue -- this time as a full
> submission (not pre-submission) and remember to fill the space where you're
> asked to point to the pre-submission.
>
